Added channel creation, moved datamodel file
